backend/services/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Authentication Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "fastapi[standard]>=0.122.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlmodel>=0.0.27",
]
[tool.pytest.ini_options]
pythonpath = ["."]
